Fix errors in the DM algo

+++ src/main/kotlin/attributes/DecisionMapUser.kt
@@ -2,4 +2,4 @@ package com.necroworld.attributes
 
 import org.hexworks.amethyst.api.Attribute
 
-data class DecisionMapUser(val weightedTags: HashMap<String, Int>) : Attribute
+data class DecisionMapUser(val weightedTags: HashMap<String, Double>) : Attribute
